fixed overlay wont hide bug
