+++ ReactQ_screens/SpendingInputReactQ.tsx
@@ -1,19 +1,26 @@
 import { View, Text, Modal, StyleSheet } from 'react-native'
 import {useContext, useState} from 'react'
 import { spending } from '../models/spending'
-import SpendingCard from '../components/Expenses/SpendingCard'
 import { OverlayContext } from '../states/context/InputOverlayContext'
 import { useStoreExpense } from '../Hooks/ReactQ'
 import { useQueryClient,QueryClient } from '@tanstack/react-query'
 import { SnackBarContext } from '../states/context/SnackBarContext'
 import { AuthContext } from '../states/context/CredentialsContext'
+import ExpenseForm from '../components/Expenses/ExpenseForm'
+import { useNavigation } from '@react-navigation/native'
+import { NativeStackNavigationProp } from '@react-navigation/native-stack'
 
+/**
+ * Simple modal that displays a form to add a new spending
+ * @returns API calls to post a new spending
+ */
 const SpendingInputReactQ = () => {
     const [imageMode,setImageMode]=useState<boolean>(false)
     const {visible,toogleOverlay}=useContext(OverlayContext);
     const {setSnackBar}=useContext(SnackBarContext)
     const {userId}=useContext(AuthContext)
     const queryClient = useQueryClient()
+    const {navigate}=useNavigation<NativeStackNavigationProp<any>>()
 
     const {mutate}=useStoreExpense({onSuccess:toogleOverlay,queryClient,onError:({response}:{response:any})=>{
         toogleOverlay()
@@ -27,6 +34,15 @@ const SpendingInputReactQ = () => {
         })
     }
 
+    const imageModeHandler=()=>{
+        //about to create a new image; navigate to places form
+        navigate('PlaceForm')
+        toogleOverlay()
+        
+    }
+
+    //this function should full control the expense form
+
 
     return (
         <Modal
@@ -34,7 +50,7 @@ const SpendingInputReactQ = () => {
             visible={visible} 
             animationType={'fade'}
             transparent={true}>
-            <SpendingCard confirm={submitAction}/>
+            <ExpenseForm confirm={submitAction} imageModeHandler={imageModeHandler} optionalButton={toogleOverlay}/>
         </Modal>
         
     )
